refactor: split util.py into text_format, title_variants, and fragment_handlers modules
Extract focused modules from library/util.py:
- library/text_format.py: ascii_text, html_text, text_with_ascii_and_emojis, text_ascii_only, path_safe_filename
- library/title_variants.py: TitleVariants class, deduplicate_variants function
- library/fragment_handlers.py: HEADING_TAGS constant, 6 fragment handler functions, _find_fragment_anchor

PageMetadata.resolve_fragment_text() now calls standalone handler functions
instead of self._fragment_handler_* methods. Backward-compatible re-exports
added to util.py so existing imports continue to work.
